refactored to use method for set/get delegated administrator instead of constructor arg
Originally I had used the opt principal in the constructor since it seemed the least invasive way to add this functionality which, in turn, at the time was also the least invasive way of setting up tests for testing the creators allowed list functionality. However its not possible to prevent a developer from accidently setting the anonymous principal as the delegated administrator in the constructor args when redeploying (without assuming they'd also make additional changes to the command used to redeploy) such that any asserts testing against this in the actor's scope would fail. Thus instead of an opt principal as a constructor arg, setting and getting was made part of the canister's API. Alternatively the tests could be refactored to use an identity created to specifically to deploy the canister, which is another task for the future; in that case no delegated administrator is needed, however the idea is assignment of such a principal would be a step towards decentralizing the invoice canister, as this principal could be actions of another canister managed by a DAO; or it could be the seller's identity who the invoice canister is being deployed for. In any case, here it was made explicitly part of the API with working tests to demonstrate its functionality as intended.